+++ src/prefect/indicators/strategy.py
@@ -4,6 +4,7 @@
 from quality.expectations.parameters import HISTORY_STRATEGY_FIELD as QUA_STRATEGY
 
 # extract
+from .extract.e1 import HISTORY_STRATEGY_FIELD as E1_STRATEGY
 from .extract.e4 import HISTORY_STRATEGY_FIELD as E4_STRATEGY
 from .extract.e5 import HISTORY_STRATEGY_FIELD as E5_STRATEGY
 
@@ -28,6 +29,7 @@ STRATEGY = {
     "i4": I4_STRATEGY,
     "i7": I7_STRATEGY,
     "t1": T1_STRATEGY,
+    "e1": E1_STRATEGY,
     "e4": E4_STRATEGY,
     "e5": E5_STRATEGY,
     "u5": U5_STRATEGY,